# IM-APP-server/docs/openapi.yaml
openapi: 3.0.3
info:
  title: IM-APP 后端增量接口
  version: 1.2.0
  description: |
    包含 OpenIM 对接、会话设置、万人转发、收藏、国家区号、群备注、用户举报中新增或实际优化过的 Go 后端接口。
    不包含未改动的普通登录、短信、用户查询、文件上传和旧 Go 对话接口。

    PostgreSQL 保存业务身份、好友、群组、转发任务状态、Kafka Outbox、同步任务和审计元数据；
    Kafka 承担万人转发的异步派发和消费，转发总人数不设业务上限；
    OpenIM + MongoDB 负责消息、会话、历史记录和 WebSocket。
servers:
  - url: http://127.0.0.1:8080
    description: 本地开发
  - url: http://8.210.72.157:8080
    description: 服务器示例（按实际部署修改）
...
paths:
...
  /api/v1/groups/{id}/mute:
    put:
      tags: [Group Sync]
      summary: 开启或关闭全员禁言
      operationId: updateGroupMute
      x-change-type: new
      security: [{bearerAuth: []}]
      parameters:
        - $ref: '#/components/parameters/GroupId'
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/UpdateGroupMuteRequest'}
      responses:
        '200': {$ref: '#/components/responses/OKResponse'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '401': {$ref: '#/components/responses/Unauthorized'}
        '403': {$ref: '#/components/responses/Forbidden'}
        '500': {$ref: '#/components/responses/InternalError'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '业务接口登录凭证，在请求头中传入 `Authorization: Bearer <accessToken>`。'
    internalApiKey:
      type: apiKey
      in: header
      name: X-Internal-API-Key
      description: 内部接口访问密钥，仅供可信后端服务和运维工具使用。
  parameters:
...
    GroupId:
      name: id
      in: path
      required: true
      description: 业务群 ID（十进制数字字符串）。
      schema: {type: string, pattern: '^[0-9]+$', example: '100001'}
...
  responses:
    OKResponse:
      description: 操作成功
      content:
        application/json:
          schema: {$ref: '#/components/schemas/OKEnvelope'}
...
    BadRequest:
      description: 参数错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
    Unauthorized:
      description: JWT 无效或缺失
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
    Forbidden:
      description: 无权限
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
    InternalError:
      description: 服务内部错误
      content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}
...
  schemas:
...
    BaseEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: {type: integer, description: 业务状态码；0 表示成功，非 0 表示失败。, example: 0}
        message: {type: string, description: 接口处理结果提示信息。, example: ok}
        data: {description: 接口返回的业务数据，具体结构见各接口响应模型。}
    ErrorEnvelope:
      type: object
      required: [code, message, data]
      properties:
        code: {type: integer, description: 错误业务码，当前失败响应通常为 1。, example: 1}
        message: {type: string, description: 可直接用于排查问题的中文错误提示。, example: 参数错误}
        data: {type: object, nullable: true, description: 失败时固定为空。}
...
    OKData:
      type: object
      required: [ok]
      properties:
        ok: {type: boolean, description: 操作是否执行成功。, example: true}
    OKEnvelope:
      allOf:
        - $ref: '#/components/schemas/BaseEnvelope'
        - type: object
          properties:
            data: {allOf: [{$ref: '#/components/schemas/OKData'}], description: 通用操作成功标记。}
...
    UpdateGroupMuteRequest:
      type: object
      required: [muted]
      properties:
        muted: {type: boolean, description: true 开启全员禁言，false 关闭全员禁言。}
...